requirements.txt
click==6.7
decorator==4.0.11
Flask==0.12.2
ipython==6.1.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.10.2
jieba==0.38
Jinja2==2.9.6
MarkupSafe==1.0
mlxtend==0.7.0
numpy==1.13.0
pandas==0.20.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
scikit-learn==0.18.2
scipy==0.19.1
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.2
